// src/app/shared/models/invoice-item.ts
export interface InvoiceItem {
  modifiedDate: number;
  itemAmt: number;
  firstName: string;
  createdDate: number;
  pickupEndtime: null;
  itemLotNum: string;
  pickedUp: boolean;
  itemDesc: string;
  lastName: string;
  lineItemId: number;
  pickupStarttime: null;
  invoiceId: number;
  itemTitle: string;
  invoiceTot: number;
  invoiceBalance: number;
  key: string;
}